Fix #90 -- Container is smaller then smallest breakpoint
An error occured, if the container was smaller then all breakpoints.
This should not happen during regular usage, and will trigger a
user warning now. However, since the API allows any container
size to be passed, this can be use the cause an internal server
error and should be handled gracefully.

Co-authored-by: truongvan <truongvan@users.noreply.github.com>